⚡️ 양도글열람페이지 ItemContent 분리
